Add files via upload
The cleaning of the data for the Ridge, Lasso, ElasticNet models is still giving me some trouble. I have tried a few preprocessing packages built into sklearn with no success(OneHotEncoder, LabelEncoder, StandardScaler). Instead, I have finally found some success with factorizing the columns that are "object" datatype, this is done on line 17 of the code. I have not had the chance to work with additional variables,  but using "Big County" and "Big Age" I have gotten the RMSE down to 23.01. You can see in the code that all these models are implemented is basically the same way. Feel free to continue tinkering with this to add more variables. I will do the same and contribute any improvements.